\name{AirPassengersX12}
\alias{AirPassengersX12}
\docType{data}
\title{
x12Single object
}
\description{
x12 Single object with the AirPassengers time series
}
\usage{data(AirPassengersX12)}
\examples{
data(AirPassengersX12)
summary(AirPassengersX12)
summary(AirPassengersX12,oldOutput=10)
}
\keyword{datasets}
