\name{AirPassengersX12Batch}
\alias{AirPassengersX12Batch}
\docType{data}
\title{
x12Batch object
}
\description{
x12Batch object of four AirPassengers series with paramters and output objects
}
\usage{data(AirPassengersX12Batch)}
\examples{
data(AirPassengersX12Batch)
summary(AirPassengersX12Batch)
}
\keyword{datasets}
