\name{PlotX12GUI}
\alias{PlotX12GUI}
\title{Plot X12 GUI}
\description{
GUI for customizing plots
}
\usage{PlotX12GUI()
}
\arguments{
  none
  }
\details{
 Each Plot is plotted clicking the Plot Button in the corresponding tab. 
}
\seealso{
  \code{\link{ts}},
  \code{\link[forecast]{forecast}},
  \code{\link{x12}}
}
\examples{
### Examples
  data(AirPassengersX12)
  #PlotX12GUI()
}
\note{
  Uses objects of class x12.
}
\author{Alexander Kowarik}