\name{spectrum-class}
\Rdversion{1.1}
\docType{class}
\alias{spectrum-class}

\title{Class \code{"spectrum"}}
\description{
Standardized object for saving the spectrum output of the x12 binaries in R. Used by functions in this package.
}
\section{Objects from the Class}{
Objects can be created by calls of the form \code{new("spectrum", ...)}.
}
\section{Slots}{
  \describe{
    \item{\code{frequency}:}{Object of class \code{"numeric"} ~~ }
    \item{\code{spectrum}:}{Object of class \code{"numeric"} ~~ }
  }
}
\author{
Alexander Kowarik
}
\examples{
showClass("spectrum")
}
\keyword{classes}
