\name{X12GUI}
\alias{X12GUI}
\title{X12 GUI}
\description{
GUI for the wrapper function for x12.
}
\usage{X12GUI()
}
\arguments{
  none
  }
\details{
 The GUI provides access to the basic parameters of the X12 method. It is based on the X12 windows interface.
 For users familiar with X12, the GUI should create a spc file as starting point for further modifications.
}
\seealso{
  \code{\link{ts}},
  \code{\link[forecast]{forecast}},
  \code{\link{x12}}
}
\examples{
### Examples
}
\note{
 Only working on with available x12 binaries.
}
\author{Alexander Kowarik}