\name{AirPassengersX12}
\alias{AirPassengersX12}
\docType{data}
\title{AirPassengers Data Output from X12}
\description{
  The dataset \code{\link{AirPassengers}}
  
}
\usage{
data(AirPassengersX12)
}
\format{
  see \code{\link{AirPassengers}}

}
\note{
    For a more detailed description of this data set, see 
    \code{\link{AirPassengers}}.
}
\references{
\code{\link{AirPassengers}}
}
\seealso{\code{\link{AirPassengers}}}
\examples{
data(AirPassengersX12)
plot(AirPassengersX12)
}
\keyword{datasets}
